requirements.txt
# Django LMS Production Requirements - Consolidated & Optimized
# All versions tested and compatible with each other
# Last updated: 2025-10-18 - Updated for Python 3.9+ compatibility

# Core Django Framework
# NOTE: Django 3.2.25 is nearing end of security support (April 2024)
# Plan migration to Django 4.2 LTS for long-term security support
Django==3.2.25
asgiref==3.9.1
sqlparse==0.4.4
pytz==2024.1

# Database Drivers
psycopg2-binary==2.9.9
dj-database-url==2.3.0

# Authentication & Security
django-allauth==0.61.1
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography==43.0.0  # Latest stable version with security updates
oauthlib==3.2.2
requests-oauthlib==2.0.0
python3-openid==3.2.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
pyotp==2.9.0

# Static Files & Media
whitenoise==6.6.0
Pillow==10.4.0
pilkit==3.0

# Web Server
gunicorn==23.0.0

# Background Tasks (Redis/Celery)
celery==5.3.6
redis==5.0.7
billiard==4.2.0
kombu==5.3.6
amqp==5.3.1
vine==5.1.0

# Content Processing
bleach==6.0.0
beautifulsoup4==4.12.2
django-crispy-forms==2.4
django-multiupload==0.6.1
soupsieve==2.4.1

# Reports & Excel Processing
reportlab==4.0.4
psutil==6.0.0  # Updated to latest stable version
openpyxl==3.1.2
XlsxWriter==3.1.9
xlwt==1.3.0
xlrd==2.0.1
pandas==1.5.3
numpy==1.24.4
python-dateutil==2.8.2
pdfplumber==0.9.0

# File Processing
python-magic==0.4.27
chardet==5.2.0

# Content Processing
Markdown==3.4.4
MarkupSafe==2.1.3

# HTTP & API
requests==2.31.0
urllib3==1.26.18  # Pinned to specific version for security and stability
djangorestframework==3.14.0

# Utilities
six==1.16.0
packaging==23.2
python-dotenv==1.0.0
humanize==4.8.0

# AWS & Storage
boto3==1.34.0
botocore==1.34.0
s3transfer==0.9.0
django-storages==1.14.1

# WebSockets & Channels
channels==3.0.5
channels-redis==4.1.0

# Development Tools
django-filter==23.3
django-widget-tweaks==1.5.0
django-import-export==3.3.1

# Type Checking and Static Analysis
djangorestframework-stubs==3.14.5
